# backend/docs/swagger.yaml
basePath: /api/
definitions:
...
  api.bookmarkInput:
    properties:
      icon:
        description: as in Icon; "" uses the linked host's or service's
        type: string
      id:
        type: integer
      mac:
        description: linked host, "" for none
        type: string
      name:
        type: string
      note:
        description: shown under the name instead of the address
        type: string
      port:
        description: linked service on that host, 0 for the host itself
        type: integer
      tags:
        items:
          type: string
        type: array
      url:
        type: string
    type: object
...
  models.Bookmark:
    properties:
      icon:
        description: as in Icon; "" uses the linked host's or service's
        type: string
      id:
        type: integer
      mac:
        description: linked host, "" for none
        type: string
      name:
        type: string
      note:
        description: shown under the name instead of the address
        type: string
      port:
        description: linked service on that host, 0 for the host itself
        type: integer
      url:
        type: string
    type: object
  models.BookmarkInfo:
    properties:
      hostID:
        description: 0 when not linked, or the host is gone
        type: integer
      hostName:
        description: linked host's name or IP
        type: string
      icon:
        description: as in Icon; "" uses the linked host's or service's
        type: string
      id:
        type: integer
      mac:
        description: linked host, "" for none
        type: string
      name:
        type: string
      note:
        description: shown under the name instead of the address
        type: string
      port:
        description: linked service on that host, 0 for the host itself
        type: integer
      tags:
        items:
          type: string
        type: array
      url:
        type: string
    type: object
...
  models.Connector:
    properties:
      enabled:
        type: boolean
      hasToken:
        type: boolean
      hostIP:
        description: 'docker: IP of the host running Docker, if not the URL''s'
        type: string
      id:
        type: integer
      insecure:
        description: accept self-signed certificates
        type: boolean
      interval:
        description: minutes between syncs
        type: integer
      kind:
        description: '"docker", "dockhand", "scanopy", "unifi", "technitium" or "proxmox"'
        type: string
      lastCount:
        description: hosts the last sync matched
        type: integer
      lastError:
        type: string
      lastSync:
        type: string
      name:
        type: string
      site:
        description: UniFi site, Dockhand environment; "" for all
        type: string
      url:
        type: string
      user:
        description: for sources that log in with a user name
        type: string
    type: object
  models.ConnectorSave:
    properties:
      clearToken:
        type: boolean
      enabled:
        type: boolean
      hasToken:
        type: boolean
      hostIP:
        description: 'docker: IP of the host running Docker, if not the URL''s'
        type: string
      id:
        type: integer
      insecure:
        description: accept self-signed certificates
        type: boolean
      interval:
        description: minutes between syncs
        type: integer
      kind:
        description: '"docker", "dockhand", "scanopy", "unifi", "technitium" or "proxmox"'
        type: string
      lastCount:
        description: hosts the last sync matched
        type: integer
      lastError:
        type: string
      lastSync:
        type: string
      name:
        type: string
      site:
        description: UniFi site, Dockhand environment; "" for all
        type: string
      token:
        type: string
      url:
        type: string
      user:
        description: for sources that log in with a user name
        type: string
    type: object
  models.Container:
    properties:
      category:
        type: string
      cid:
        description: container ID
        type: string
      connectorID:
        type: integer
      host:
        type: string
      icon:
        description: icon suggestion from the image name
        type: string
      id:
        type: integer
      image:
        type: string
      ip:
        description: IP of the Docker host
        type: string
      mac:
        description: host it runs on, "" if that host is not known
        type: string
      name:
        type: string
      ports:
        items:
          $ref: '#/definitions/models.ContainerPort'
        type: array
      project:
        description: compose project
        type: string
      state:
        description: '"running", "exited"...'
        type: string
      status:
        type: string
    type: object
  models.ContainerPort:
    properties:
      ip:
        type: string
      private:
        type: integer
      proto:
        type: string
      public:
        description: 0 when not published on the host
        type: integer
    type: object
...
info:
  contact:
    url: https://github.com/vipthomps/Gimle
  description: 'Home lab dashboard: network discovery, a map of hosts, bookmarks,
    views and docs'
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Gimlé API
  version: 1.0.0
paths:
...
  /backups/{name}/restore:
    post:
      description: Puts back the database and settings, keeping the address and port
        in use. The current state is backed up first, so a restore can be undone.
      parameters:
      - description: Backup file name
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Restore a backup
      tags:
      - backups
  /bookmarks:
    get:
      description: Addresses added by hand, with their tags and the host they are
        linked to
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.BookmarkInfo'
            type: array
      summary: List bookmarks
      tags:
      - bookmarks
    post:
      consumes:
      - application/json
      description: ID 0 creates one. URL must be http(s). Mac links it to a host,
        Port to one of its services. Tags replace the bookmark's tags.
      parameters:
      - description: Bookmark with tags
        in: body
        name: bookmark
        required: true
        schema:
          $ref: '#/definitions/api.bookmarkInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Bookmark'
      summary: Create or update a bookmark
      tags:
      - bookmarks
  /bookmarks/{id}:
    delete:
      description: Also removes it from every group it was tagged into
      parameters:
      - description: Bookmark ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete a bookmark
      tags:
      - bookmarks
...
  /connectors:
    get:
      description: Docker, Dockhand, Scanopy, UniFi and Technitium sources. Tokens
        are never returned.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Connector'
            type: array
      summary: List connectors
      tags:
      - connectors
    post:
      consumes:
      - application/json
      description: ID 0 creates one. An empty Token keeps the saved token; ClearToken
        removes it.
      parameters:
      - description: Connector
        in: body
        name: connector
        required: true
        schema:
          $ref: '#/definitions/models.ConnectorSave'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Connector'
      summary: Create or update a connector
      tags:
      - connectors
  /connectors/{id}:
    delete:
      description: Also removes the containers and ports it reported
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: Delete a connector
      tags:
      - connectors
  /connectors/{id}/sync:
    post:
      description: Waits for the sync and returns the connector with its result
      parameters:
      - description: Connector ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Connector'
      summary: Sync a connector now
      tags:
      - connectors
...
  /containers:
    get:
      description: Containers the connectors reported, optionally for one host
      parameters:
      - description: Host MAC
        in: query
        name: mac
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/models.Container'
            type: array
      summary: Containers
      tags:
      - connectors
...
swagger: "2.0"